+++ internal/cmd/create.go
@@ -15,8 +15,8 @@ func newCreateCommand(cmdCfg *zsmCommandConfig) *cobra.Command {
 If [FILE SYSTEM] is passed, then a snapshot of only the passed file system is
 created unless [FILE SYSTEM] is marked as excluded.
 
-The created snapshots start with the same name as the dataset and are suffixed with @<TIMESTAMP>
-where <TIMESTAMP> is an RFC3339 timestamp. The time zone of the <TIMESTAMP> is always UTC regardles
+The created snapshots start with the same name as the dataset and are suffixed with @TIMESTAMP
+where TIMESTAMP is an RFC3339 timestamp. The time zone of the TIMESTAMP is always UTC regardles
 of the system time.`,
 		Args: cobra.RangeArgs(0, 1),
 		RunE: func(cmd *cobra.Command, args []string) error {